add feature realTime.

+++ views/app/controllers/roomUserController.js
@@ -28,7 +28,7 @@ class roomUserController {
     condition = {
       ...condition, ...res.query
     }
-    
+
     roomUserModel.find(condition).exec((err, data) => {
       if (err) {
         return res.json({
@@ -36,11 +36,11 @@ class roomUserController {
           status: statusF,
           data: [],
         });
-      }else if(data.length === 0){
+      } else if (data.length === 0) {
         return res.json({
-            status: statusF,
-            data: data,
-            message: "Không tìm thấy dữ liệu theo yêu cầu."
+          status: statusF,
+          data: data,
+          message: "Không tìm thấy dữ liệu theo yêu cầu."
         })
       } else {
         return res.json({
@@ -64,12 +64,12 @@ class roomUserController {
           data: [],
           message: `We have some error:${resp}`
         })
-      }else if(resp.length === 0){
-          return res.json({
-              status: statusF,
-              data: resp,
-              message: "Không tìm thấy phòng yêu cầu.",
-          })
+      } else if (resp.length === 0) {
+        return res.json({
+          status: statusF,
+          data: resp,
+          message: "Không tìm thấy phòng yêu cầu.",
+        })
       } else {
         return res.json({
           status: statusS,
@@ -83,24 +83,33 @@ class roomUserController {
 
     try {
       let { id_Room: idR, id_User: idU } = req.body;
-      
-      if(!idR || !idU){
-          return res.json({
-              status: statusF,
-              data: [],
-              message: "Vui lòng nhập đủ thông tin."
-          })
+
+      const checkExistUser = await roomUserModel.find({ id_User: idU });
+      if (checkExistUser.length) {
+        return res.json({
+          status: statusS,
+          data: checkExistUser,
+          message: "user was existed on room."
+        })
       }
 
-    //   let findRoom = await song.find({ _id: mongoose.Types.ObjectId(idR) });
-      let findUser = await userModel.find({ _id: mongoose.Types.ObjectId(idU) });
+      if (!idR || !idU) {
+        return res.json({
+          status: statusF,
+          data: [],
+          message: "Vui lòng nhập đủ thông tin."
+        })
+      }
 
-    if (findUser.length !== 0) {
-    // if (findRoom.length !== 0 || findUser.length !== 0) {
+      //   let findRoom = await song.find({ _id: mongoose.Types.ObjectId(idR) });
+      const findUser = await userModel.find({ _id: idU });
+
+      if (findUser.length !== 0) {
+        // if (findRoom.length !== 0 || findUser.length !== 0) {
 
         let newData = {
-          id_Room: mongoose.Types.ObjectId(idR),
-          id_User: mongoose.Types.ObjectId(idU),
+          id_Room: idR,
+          id_User: idU,
         }
 
         let createRoomUser = await new roomUserModel(newData)
@@ -118,15 +127,16 @@ class roomUserController {
             })
           }
         });
-      }else{
-          return res.json({
-              status: statusF,
-              data: [],
-              message: "Tài khoản hoặc Phòng không tồn tại, User or Room does not exist."
-          });
+      } else {
+        return res.json({
+          status: statusF,
+          data: [],
+          message: "Tài khoản hoặc Phòng không tồn tại, User or Room does not exist."
+        });
       }
 
     } catch (error) {
+      console.log(error)
       res.json({
         status: statusF,
         data: [],
@@ -136,54 +146,54 @@ class roomUserController {
   }
   async update(req, res) {
     try {
-        let { id_Room: idR, id_User: idU } = req.body;
+      let { id_Room: idR, id_User: idU } = req.body;
 
-        if(!idR || !idU){
-            return res.json({
-                status: statusF,
-                data: [],
-                messgae: "Vui lòng nhập đủ thông tin.",
-            })
-        };
+      if (!idR || !idU) {
+        return res.json({
+          status: statusF,
+          data: [],
+          messgae: "Vui lòng nhập đủ thông tin.",
+        })
+      };
 
-        // let findRoom = await song.find({ _id: mongoose.Types.ObjectId(idR) });
-        let findUser = await userModel.find({ _id: mongoose.Types.ObjectId(idU) });
+      // let findRoom = await song.find({ _id: mongoose.Types.ObjectId(idR) });
+      let findUser = await userModel.find({ _id: mongoose.Types.ObjectId(idU) });
 
-        if (findUser.length !== 0) {
+      if (findUser.length !== 0) {
         // if (findRoom.length !== 0 || findUser.length !== 0) {
-            let newData = {
-                id_Room: mongoose.Types.ObjectId(idR),
-                id_User: mongoose.Types.ObjectId(idR),
-            }
-
-            let idRoomUser = req.params.idRoomUser;
-
-            const condition = {
-                _id: mongoose.Types.ObjectId(idRoomUser)
-            }
-
-            roomUserModel.findOneAndUpdate(condition, { $set: newData }, { new: true })
-            .exec((err, new_data) => {
-                if (err) {
-                    return res.json({
-                        status: "failed",
-                        message: `We have few error: ${err}`
-                    })
-                } else {
-                    return res.json({
-                        status: "successfully",
-                        data: [new_data],
-                        message: `You were update successfully`
-                    })
-                }
-            })
-        }else{
-            return res.json({
-                status: statusF,
-                data: [],
-                message: "Bài hát hoặc Tài khoản không tồn tại, Song or User does not exist."
-            });
+        let newData = {
+          id_Room: mongoose.Types.ObjectId(idR),
+          id_User: mongoose.Types.ObjectId(idR),
         }
+
+        let idRoomUser = req.params.idRoomUser;
+
+        const condition = {
+          _id: mongoose.Types.ObjectId(idRoomUser)
+        }
+
+        roomUserModel.findOneAndUpdate(condition, { $set: newData }, { new: true })
+          .exec((err, new_data) => {
+            if (err) {
+              return res.json({
+                status: "failed",
+                message: `We have few error: ${err}`
+              })
+            } else {
+              return res.json({
+                status: "successfully",
+                data: [new_data],
+                message: `You were update successfully`
+              })
+            }
+          })
+      } else {
+        return res.json({
+          status: statusF,
+          data: [],
+          message: "Bài hát hoặc Tài khoản không tồn tại, Song or User does not exist."
+        });
+      }
     } catch (error) {
       return res.json({
         status: statusF,
@@ -194,7 +204,10 @@ class roomUserController {
   }
   delete(req, res) {
     const condition = {
-      _id: mongoose.Types.ObjectId(req.params.idRoomUser)
+      _id: req.params.idRoomUser
+    }
+    if (!condition._id) {
+      return res.json({ status: statusF })
     }
     roomUserModel.findOneAndRemove(condition)
       .exec((err) => {
